Refactor determining the handler for a node id call
