+++ command/agent/stats_endpoint.go
@@ -12,26 +12,14 @@ func (s *HTTPServer) ClientStatsRequest(resp http.ResponseWriter, req *http.Requ
 	// Get the requested Node ID
 	requestedNode := req.URL.Query().Get("node_id")
 
-	localClient := s.agent.Client()
-	localServer := s.agent.Server()
-
-	// See if the local client can handle the request.
-	useLocalClient := localClient != nil && // Must have a client
-		(requestedNode == "" || // If no node ID is given
-			requestedNode == localClient.NodeID()) // Requested node is the local node.
-
-	// Only use the client RPC to server if we don't have a server and the local
-	// client can't handle the call.
-	useClientRPC := localClient != nil && !useLocalClient && localServer == nil
-
-	// Use the server as a last case.
-	useServerRPC := localServer != nil && requestedNode != ""
-
 	// Build the request and parse the ACL token
 	args := cstructs.ClientStatsRequest{
 		NodeID: requestedNode,
 	}
 	s.parse(resp, req, &args.QueryOptions.Region, &args.QueryOptions)
+
+	// Determine the handler to use
+	useLocalClient, useClientRPC, useServerRPC := s.rpcHandlerForNode(requestedNode)
 
 	// Make the RPC
 	var reply cstructs.ClientStatsResponse
